+++ src/main/java/com/rakovets/course/java/core/practice/date_and_time/DateWrapper.java
@@ -33,19 +33,4 @@ public abstract class DateWrapper {
     public static long getDaysBetweenDates(LocalDate first, LocalDate second) {
         return Math.abs(DAYS.between(first, second));
     }
-
-    public static LocalDate getDateAfterDays (LocalDate date, int days) {
-        TemporalAdjuster afterDays = temporal -> temporal.plus(days, DAYS);
-        return (LocalDate) afterDays.adjustInto(date);
-    }
-
-    public static TemporalAdjuster closestNewYear = temporal ->
-            (temporal.get(ChronoField.DAY_OF_YEAR) >= (temporal.range(ChronoField.DAY_OF_YEAR).getMaximum() / 2))
-                    ? temporal.with(TemporalAdjusters.firstDayOfNextYear())
-                    : temporal.with(TemporalAdjusters.firstDayOfYear());
-
-    public static Date getDate(int year, int month, int days) throws ParseException {
-        String string = days + "." + month + "." + year;
-        return new SimpleDateFormat("dd.MM.yyyy").parse(string);
-    }
 }
